simulation/requirements.txt
fastapi
httpx
uvicorn
uvicorn[standard]
PyYAML
numpy<2
PILLOW
opencv-python
catkin-pkg
